Switch send back count to 16 bits

+++ qt_quick_app/scanner.h
@@ -13,7 +13,7 @@ class Scanner : public AFMObject
     Q_OBJECT
     Q_PROPERTY(quint8 m_num_averages READ num_averages WRITE set_num_averages NOTIFY num_averages_changed)
     Q_PROPERTY(quint8 m_dwell_time READ dwell_time WRITE set_dwell_time NOTIFY dwell_time_changed)
-    Q_PROPERTY(quint8 m_send_back_count READ send_back_count WRITE set_send_back_count NOTIFY send_back_count_changed)
+    Q_PROPERTY(quint16 m_send_back_count READ send_back_count WRITE set_send_back_count NOTIFY send_back_count_changed)
     Q_PROPERTY(quint16 m_num_rows READ num_rows WRITE set_num_rows NOTIFY num_rows_changed)
     Q_PROPERTY(quint16 m_num_columns READ num_columns WRITE set_num_columns NOTIFY num_columns_changed)
     Q_PROPERTY(double m_rms_threshold READ rms_threshold WRITE set_rms_threshold NOTIFY rms_threshold_changed)
@@ -31,7 +31,7 @@ public:
     Q_INVOKABLE void set_num_averages(int num_averages);
     Q_INVOKABLE quint8 dwell_time();
     Q_INVOKABLE void set_dwell_time(int dwell_time);
-    Q_INVOKABLE quint8 send_back_count();
+    Q_INVOKABLE quint16 send_back_count();
     Q_INVOKABLE void set_send_back_count(int send_back_count);
     Q_INVOKABLE quint16 num_rows();
     Q_INVOKABLE void set_num_rows(int num_rows);
@@ -89,7 +89,7 @@ private:
     quint8 m_ratio;
     quint8 m_dwell_time;
     quint8 m_num_averages;
-    quint8 m_send_back_count;
+    quint16 m_send_back_count;
     double m_rms_threshold;
 
     int m_num_columns_received;
